Tratamento de Erros

+++ src/main/java/com/example/demo/services/UserService.java
@@ -4,6 +4,7 @@ import com.example.demo.entities.User;
 import com.example.demo.entities.dto.ResponseUserDTO;
 import com.example.demo.entities.dto.UserDTO;
 import com.example.demo.entities.enums.RoleUser;
+import com.example.demo.exeptions.notfound.UserNotFoundException;
 import com.example.demo.repositories.UserRepository;
 import com.example.demo.security.UserDetailsServiceImpl;
 import com.example.demo.security.jwt.JWTService;
@@ -33,8 +34,9 @@ public class UserService {
         return repository.findAll(pageable).map(user -> new ResponseUserDTO(user));
     }
 
-    public ResponseUserDTO findUserById(Long id) {
-        User user = repository.findById(id).get();
+    public ResponseUserDTO findUserById(Long id) throws Exception {
+        User user = repository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("Usuário Não Encntrado!"));
         return new ResponseUserDTO(user);
     }
 
@@ -44,12 +46,14 @@ public class UserService {
         return new ResponseUserDTO(repository.save(user));
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws Exception {
+        if(repository.existsById(id) == false) throw new UserNotFoundException("Usuário Não Encntrado!");
         repository.deleteById(id);
     }
 
-    public ResponseUserDTO update(Long id, UserDTO userDTO) {
-        User entity = repository.findById(id).get();
+    public ResponseUserDTO update(Long id, UserDTO userDTO) throws Exception {
+        User entity = repository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("Usuário Não Encntrado!"));
         updateData(entity, userDTO);
         return new ResponseUserDTO(repository.save(entity));
     }
@@ -60,10 +64,11 @@ public class UserService {
 
         try {
             if (senhaEhValida) {
-                User usuarioDoBanco = repository.findUserByUsername(user.getUsername());
+                User usuarioDoBanco = repository.findUserByUsername(user.getUsername())
+                        .orElseThrow(() -> new UserNotFoundException("Username Incorreto!"));
                 return new TokenDTO(Jwt.generationToken(usuarioDoBanco));
             }
-        } catch (UsernameNotFoundException e) {
+        } catch (UsernameNotFoundException | UserNotFoundException e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Username ou Senha Inválidos!");
         }
         throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Senha Inválida!");
